Incorporate runner into Docker

+++ scripts/run.py
@@ -1,7 +1,10 @@
+#!/usr/bin/env python3
 import subprocess
-from concurrent.futures import ProcessPoolExecutor
+from multiprocessing import Pool
+from tqdm import tqdm
 from pathlib import Path
 import argparse
+import random
 
 
 def parse_args():
@@ -10,14 +13,20 @@ def parse_args():
     parser.add_argument("-o", "--output", required=True, type=Path)
     parser.add_argument("-b", "--binary", required=True, type=Path)
     parser.add_argument("-n", "--num_threads", type=int, default=4)
+    parser.add_argument("-s", "--skip_existing", action="store_true")
     return parser.parse_args()
 
 
 def process_file(args):
-    input_file, output_dir, binary_path = args
-    output_file = output_dir / input_file.stem
+    input_file, output_dir, binary_path, skip_existing = args
+    name = input_file.stem
+    if name.endswith(".gr"):
+        name = name[:-3]
 
-    print(f"Run on {input_file}")
+    output_file = output_dir / f"{name}.log"
+    if skip_existing and output_file.exists():
+        return
+
     with open(output_file, "w") as out:
         proc = subprocess.Popen(
             [binary_path, "--file", input_file],
@@ -29,12 +38,12 @@ def process_file(args):
 def main():
     args = parse_args()
     args.output.mkdir(parents=True, exist_ok=True)
-    files = list(args.input.iterdir())
-    job_args = [(f, args.output, args.binary) for f in files]
+    files = list(args.input.glob("*.gr")) + list(args.input.glob("*.gr.bz2"))
+    random.shuffle(files)
+    job_args = [(f, args.output, args.binary, args.skip_existing) for f in files]
 
-    with ProcessPoolExecutor(max_workers=args.num_threads) as executor:
-        executor.map(process_file, job_args)
-
+    with Pool(args.num_threads) as pool:
+        list(tqdm(pool.imap_unordered(process_file, job_args, chunksize=1), total=len(job_args)))
 
 if __name__ == "__main__":
     main()
